+++ tests/testthat/test-tbl_likert.R
@@ -115,3 +115,20 @@ test_that("tbl_likert(sort)", {
       as.data.frame()
   )
 })
+
+# addressing issue # 2195
+# check the order of the variables match the input, and the formatting of the N
+test_that("tbl_likert(sort)", {
+  expect_snapshot(
+    withr::with_seed(
+      seed = 11235,
+      data.frame(
+        recommend_friend = sample(levels, size = 1001, replace = TRUE) |> factor(levels = levels),
+        regret_purchase = sample(levels, size = 1001, replace = TRUE) |> factor(levels = levels)
+      )
+    ) |>
+      tbl_likert(include = c(regret_purchase, recommend_friend)) |>
+      add_n() |>
+      as.data.frame()
+  )
+})